+++ src/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import ResponsiveAppBar from "./components/AppBar";
+import ResponsiveAppBar from "./components/AppBar/AppBar";
 import Home from "./views/Home/Home";
 import Kontakt from "./views/Kontakt/Kontakt";
 import { Routes, Route } from "react-router-dom";
